Extensions installation functionality for Azure CLI.
For some rules to behave as expected, you need to ensure that some extensions are available in the Azure CLI.
A feature has been created that supports the definition of extensions that must be installed.
During the loading phase of the repository, it is guaranteed that the
extensions will be installed and made available for use in the rules.
Some adjustments and improvements need to be made, such as validating whether the extension is available in the build/run phase.