Support for the az_config rule call in the repository analysis phase.
The az_config rule aims to define basic properties needed to interact with Azure.
Creating az_config rules at the workspace level helps to keep organized settings that must be global, as well as avoid repetition and inconsistency in relation to defined attributes.

The rule definition in the WORKSPACE file is just an alias for the rule that can also be referenced in BUILD.bazel files.
So it is still possible to call the rule in the BUILD.bazel file.