Added artifacts for building and creating stamping files
To create stamping files, we are creating a program to run in the build phase to read variables from the workspace.
This program was done in GoLang. The executable is created from the rules_go rules of bazel.
The rules_gazelle rule was also used to automate the process of creating BUILD.bazel, defining dependencies, among other facilities.